refactor: 프로그래밍 요구 사항에 따른 Lotto 클래스 파일 원위치
